Committing first version of mixminion code to CVS repository.  Right now, the code to generate and handle messages is in place, but we're still missing all the networking logic... along with everything else mentioned in TODO.
See HACKING if you want to make it go.

See TODO it you want to make it better.

Please don't circulate this beyond the list until we have more of it
working.

I'm going to have to work on my day job for a while, so I may be slow
writing more of this.  Nevertheless, I hope to get the
buildMessage/serverProcess logic well tested and debugged by this
weekend.