More tests for everything.
BuildMessage and ServerProcess are now working: we support all of the
non-network-related algorithms and formats in the spec.  We can build
messages and reply blocks, and (if we're a server) process them enough
to pass messages off to the next server.  We handle extended headers,
swapping, everything.

Performance is good: on my Athlon XP1700+, the server core takes 8.5
ms for a non-swap message, and 9 ms for a swap message.  Of this, I
believe around 95% is spent in the C encryption code.

Next up, more documentation and comments and code clean-up.  After
that, get queues and persistent server information working.