Fix a nasty bug in the scheduler code.
If any scheduled event (shred, timeout, or mix) took so long to complete
that the next event would start immediately, the next event would reschedule
itself based on the time of the *first* event.  If the first event took
longer than the reschedule interval of the second event, the second event
would reschedule itself in the past, and start the cycle over again.