Fix error in path-length checking.
This was a weird one.  Remember how a path contains:

    RSA_ENCRYPTED SYM_ENCRYPTED_______________________
    datadata RSA_ENCRYPTED SYM_ENCRYPTED______________
             datadata RSA_ENCRYPTED SYM_ENCRYPTED_____
                   ......
                                       RSA_ENC SYM_ENC
                                       datadata padding  ?

Previously, if all of the 'datadata' chunks were big enough to fit
into a header, we were happy.  But there was an error case where all
of the 'datadata' chunks fit, but there isn't enough room to pad the
_final_ RSA-encrypted block up to 256 bytes.  Now we check for this
case too.