Stopgap fix to earlier bugfix.
It seems that the *method() calls in OpenSSL are not as versitile as I
hoped, and that a TLSv1 server (all existing servers!) will not talk
to a SSL23 client (oops!).

[SSL23 means "connect with ssl2, ssl3, or tls1."  We disable SSL2.]

The right solution is to make all code continue to use TLSv1 for
clients, and use SSL23 for servers (which the clients can force into
TLSv1 more).  With 0.0.7, we can make the clients go SSL23 as well.