Begin migrating to support the spec's alternative SSL crypto suite.
The spec says that we should support an alternative (and more common)
crypto suite for client-to-server communications.  The alternative
suite is only present in SSL3; the preferred one is in TLS1.

Older versions of the code are configured to generate only TLS1
connections -- and (previously unknown to me) accept only TLS1
connections.  To do the right thing, we need to accept TLS1 and SSL3,
but generate only TLS1.  This patch does that.