Removed CVS-Id-tags and for-emacs-lines
Each file in the repository had a CVS-Id-tag on top. As we now moved to git
this is not needed anymore. However the file contrib/textwrap.py also had the
line:
  __revision__ = "$Id: textwrap.py,v 1.2 2003/06/21 21:44:01 nickm Exp $"
This was also removed.

Furthermore some files had some settings for GNU/Emacs. I also consider them as not necessary and removed it.